// packages/community/src/types/SignUpInputType.ts
import InputFormatType from './InputFormatType';

interface SignUpInputType {
  id: InputFormatType;
  password: InputFormatType;
  passwordCf: InputFormatType;
  username: InputFormatType;
  email: InputFormatType;
  mobile: InputFormatType;
  aaaNum: InputFormatType;
  schoolNum: InputFormatType;
  major: InputFormatType;
  introduction: InputFormatType;
}

export default SignUpInputType;